moving more stuff into the shared server package
