+++ apps/zupoll-server/package.json
@@ -22,6 +22,7 @@
     "@pcd/semaphore-group-pcd": "0.11.0",
     "@pcd/semaphore-signature-pcd": "0.11.0",
     "@pcd/util": "^0.5.0",
+    "@pcd/server-shared": "0.0.1",
     "@pcd/zupoll-shared": "0.0.1",
     "@prisma/client": "^4.12.0",
     "cors": "^2.8.5",
