+++ apps/zupoll-server/src/application.ts
@@ -1,3 +1,4 @@
+import { RollbarService, getCommitHash, requireEnv } from "@pcd/server-shared";
 import { Bot } from "grammy";
 import { startBot } from "./bot";
 import { startServer } from "./routing/server";
@@ -5,7 +6,8 @@ import { logger } from "./util/logger";
 
 export async function startApplication() {
   const context: ApplicationContext = {
-    bot: undefined
+    bot: undefined,
+    rollbarService: await startRollbarService()
   };
 
   startBot(context).catch((e) => {
@@ -17,4 +19,29 @@ export async function startApplication() {
 
 export interface ApplicationContext {
   bot?: Bot;
+  rollbarService?: RollbarService;
+}
+
+export async function startRollbarService(): Promise<
+  RollbarService | undefined
+> {
+  let rollbarToken: string;
+  let rollbarEnvironmentName: string;
+
+  try {
+    rollbarToken = requireEnv("ROLLBAR_TOKEN");
+    rollbarEnvironmentName = requireEnv("ROLLBAR_ENV_NAME");
+  } catch (e) {
+    console.log(`[ROLLBAR] not starting, missing env ${e}`);
+    return undefined;
+  }
+
+  console.log(`[ROLLBAR] starting`);
+  const rollbarService = new RollbarService(
+    rollbarToken,
+    rollbarEnvironmentName,
+    await getCommitHash()
+  );
+
+  return rollbarService;
 }
