+++ packages/lib/server-shared/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./services/RollbarService";
+export * from "./util";
